refactor(mstp): split frame tests to satisfy 700 LOC cap
Move Clause 9 golden / stream tests into mstp_frame_tests.rs so
mstp_frame.rs stays under the upstream file-size CI gate.
